# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports/campaign/{campaignId}": {
      "get": {
        "operationId": "ReportsController_getCampaignReport",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "today",
                "week",
                "month",
                "quarter",
                "lifetime"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/driver-activity/rickshaw/{rickshawId}": {
      "get": {
        "operationId": "DriverActivityController_findByRickshaw",
        "parameters": [
          {
            "name": "rickshawId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverActivity"
        ]
      }
    },
...
    "/rickshaws": {
      "post": {
        "operationId": "RickshawsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRickshawDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rickshaws"
        ]
      },
      "get": {
        "operationId": "RickshawsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rickshaws"
        ]
      }
    },
    "/rickshaws/{id}": {
      "get": {
        "operationId": "RickshawsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rickshaws"
        ]
      },
      "patch": {
        "operationId": "RickshawsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRickshawDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rickshaws"
        ]
      },
      "delete": {
        "operationId": "RickshawsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rickshaws"
        ]
      }
    },
...
    "/driver-payout/rickshaw/{rickshawId}": {
      "get": {
        "operationId": "DriverPayoutController_findByRickshaw",
        "parameters": [
          {
            "name": "rickshawId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverPayout"
        ]
      }
    },
...
    "/media-assets/upload": {
      "post": {
        "operationId": "MediaAssetsController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MediaAssets"
        ]
      }
    },
    "/media-assets/upload/initiate": {
      "post": {
        "operationId": "MediaAssetsController_initiateUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateMediaUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MediaAssets"
        ]
      }
    },
    "/media-assets/upload/complete": {
      "post": {
        "operationId": "MediaAssetsController_completeUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteMediaUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MediaAssets"
        ]
      }
    },
    "/media-assets": {
      "get": {
        "operationId": "MediaAssetsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by media type",
            "schema": {
              "enum": [
                "image",
                "video"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all media assets",
        "tags": [
          "MediaAssets"
        ]
      }
    },
    "/media-assets/{id}": {
      "get": {
        "operationId": "MediaAssetsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MediaAssets"
        ]
      },
      "delete": {
        "operationId": "MediaAssetsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MediaAssets"
        ]
      }
    },
    "/player/me": {
      "get": {
        "operationId": "PlayerRuntimeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      }
    },
    "/player/sync": {
      "get": {
        "operationId": "PlayerRuntimeController_sync",
        "parameters": [
          {
            "name": "currentManifestVersion",
            "required": false,
            "in": "query",
            "description": "Current manifest version cached on the device",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      },
      "post": {
        "operationId": "PlayerRuntimeController_syncUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSyncDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      }
    },
    "/player/manifest": {
      "get": {
        "operationId": "PlayerRuntimeController_getManifest",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      }
    },
    "/player/media/{mediaId}": {
      "get": {
        "operationId": "PlayerRuntimeController_downloadMedia",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      }
    },
    "/player/heartbeat": {
      "post": {
        "operationId": "PlayerRuntimeController_heartbeat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      }
    },
    "/player/manifest/ack": {
      "post": {
        "operationId": "PlayerRuntimeController_acknowledgeManifest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestAckDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "player-api-key": []
          }
        ],
        "tags": [
          "PlayerRuntime"
        ]
      }
    },
...
  },
  "info": {
    "title": "AdOnTheGo API",
    "description": "AdOnTheGo Backend APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Enter: ApiKey <your-api-key>"
      }
    },
    "schemas": {
...
      "CreateRickshawDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateRickshawDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateMediaAssetDto": {
        "type": "object",
        "properties": {
          "imageDuration": {
            "type": "number",
            "description": "Duration in seconds for images only",
            "example": 10
          }
        }
      },
      "InitiateMediaUploadDto": {
        "type": "object",
        "properties": {}
      },
      "CompleteMediaUploadDto": {
        "type": "object",
        "properties": {}
      },
      "HeartbeatDto": {
        "type": "object",
        "properties": {}
      },
      "ManifestAckDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSyncDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}